Restructure directories and packages
Move all cache implementations under `./cache`.
