+++ cache/memcached/memcached.go
@@ -1,4 +1,4 @@
-package memcache
+package memcached
 
 import (
 	"errors"
