+++ zedcache_test.go
@@ -16,7 +16,7 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 
-	"github.com/connylabs/zedcache/go-cache"
+	"github.com/connylabs/zedcache/cache/go-cache"
 )
 
 func TestCheckPermission(t *testing.T) {
